% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRISKcont2.R
\name{plotRISKcont2}
\alias{plotRISKcont2}
\title{Plot relative transmission probabilities across contact networks}
\usage{
plotRISKcont2(
  x,
  ref = NULL,
  Nsim = NULL,
  colors = c(SIN = "magenta3", ASN = "red2", MPN = "dodgerblue", RCN = "green3")
)
}
\arguments{
\item{x}{A `simPOP` object containing contact-network inference results in
the `RISKcont` slot, generated using `RISKcont()`.}

\item{ref}{A character string specifying the contact network used as
reference. Default is last.}

\item{Nsim}{An optional character string used as the plot title.
Default is `NULL`.}

\item{colors}{A named character vector specifying the colours used for the
candidate contact networks. Names must correspond to the contact-network
names displayed in the plot.}
}
\value{
A `ggplot` object showing the distribution of transmission
probabilities relative to the selected reference network.
}
\description{
`plotRISKcont2()` compares the transmission probabilities estimated for
candidate contact networks relative to a reference network.

For each observed infection event, the transmission probability estimated
for each candidate contact network is divided by the probability obtained
from the reference network. Values greater than `1` indicate that the
candidate network provides a higher transmission probability than the
reference network, whereas values below `1` indicate a lower probability.
}
\examples{
\dontrun{
colors=c("SIN"="magenta3", "ASN"="red2",
         "MPN"="dodgerblue", "RCN"="green3")

plotRISKcont2(POP, ref="Uniform", colors=colors)
}
}
