% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNet-methods.R
\name{show,simPOP-method}
\alias{show,simPOP-method}
\title{Show method for simPOP}
\usage{
\S4method{show}{simPOP}(object)
}
\arguments{
\item{object}{An `simPOP` object.}
}
\description{
Displays a summary of the information contained in an `simPOP` object.
}
