% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePOP.R
\name{simplePOP}
\alias{simplePOP}
\title{Simulate a simple population}
\usage{
simplePOP(NIDs = 200, steps = 100, rep = 100)
}
\arguments{
\item{NIDs}{A positive integer indicating the number of individuals in the
simulated population (Default: *NIDs=200*).}

\item{steps}{A positive integer indicating the number of transmission steps
to simulate (Default: *steps=100*).}

\item{rep}{A positive integer indicating the number of simulation replicates
(Default: *rep=100*).}
}
\value{
A `simPOP` object containing the simulated population. The
`statusM` slot contains one status matrix per replicate, with individuals in
rows and simulation steps in columns. All individuals are initially coded as
susceptible (`0`).
}
\description{
`simplePOP` creates a simple population for stochastic transmission
simulations. The population contains a specified number of individuals,
simulation steps, and replicates, without additional individual-level
information.

All individuals are initially assigned as susceptible. Contact networks and
transmission parameters can subsequently be added using other `SimNet`
functions.
}
\examples{
# Create a population of 200 individuals
POP=simplePOP(NIDs = 200, steps = 100, rep = 100)
POP

}
